Refactoring of C code.
