+++ src/nloptr.c
@@ -45,20 +45,9 @@
 // void* f_data);
 
 #include "nloptr.h"
+#include "parsers.h"
+#include <R.h>
 #include <stdbool.h>
-
-SEXP getListElement(SEXP list, char *str) {
-  SEXP elmt = R_NilValue, names = getAttrib(list, R_NamesSymbol);
-  PROTECT(names);
-  for (size_t i = 0; i < length(list); i++) {
-    if (strcmp(CHAR(STRING_ELT(names, i)), str) == 0) {
-      elmt = VECTOR_ELT(list, i);
-      break;
-    }
-  }
-  UNPROTECT(1);
-  return elmt;
-}
 
 // The algtable table must be in sorted order for bsearch to work properly.
 ALGPAIR algtable[] = {
@@ -509,34 +498,6 @@ void func_constraints_eq(unsigned m, double *constraints, unsigned n,
   }
 
   UNPROTECT(2);
-}
-
-int parse_integer_option(SEXP R_options, char *name) {
-  SEXP R_value = PROTECT(getListElement(R_options, name));
-  int value = asInteger(R_value);
-  UNPROTECT(1);
-  return value;
-}
-
-double parse_real_option(SEXP R_options, char *name) {
-  SEXP R_value = PROTECT(getListElement(R_options, name));
-  double value = asReal(R_value);
-  UNPROTECT(1);
-  return value;
-}
-
-unsigned int parse_vector_length_option(SEXP R_options, char *name) {
-  SEXP R_value = PROTECT(getListElement(R_options, name));
-  unsigned int n = length(R_value);
-  UNPROTECT(1);
-  return n;
-}
-
-double *parse_real_vector_option(SEXP R_options, char *name) {
-  SEXP R_value = PROTECT(getListElement(R_options, name));
-  double *value = REAL(R_value);
-  UNPROTECT(1);
-  return value;
 }
 
 nlopt_opt getOptions(SEXP R_options, int num_controls,
